requirements.txt
django>=5.0
django-environ
pillow
psycopg2-binary
stripe
djangorestframework
gunicorn
